requirements.txt
asgiref==3.7.2
django==4.2.11
gunicorn==22.0.0
channels==3.0.4
django-cors-headers==4.3.1
whitenoise==6.5.0
djangorestframework==3.15.2
django-filter==22.1
psycopg2-binary==2.9.9
celery==5.3.6
django-redis==5.4.0
python-dotenv==1.0.1
pytz==2023.3.post1
sqlparse==0.5.0
typing_extensions==4.8.0
tzdata==2023.3
drf-yasg==1.21.7
boto3==1.33.10
django-storages==1.14.2
djangorestframework-simplejwt==5.3.1
channels==3.0.4
channels-redis==4.2.0
redis==5.0.0b2
beautifulsoup4==4.12.3
requests==2.32.0
lxml==5.2.1
weasyprint==52.0
pydyf==0.10.0
schedule==1.2.2
pika
